+++ lib/hayiqu.dart
@@ -19,6 +19,7 @@ export 'package:cached_network_image/cached_network_image.dart';
 export 'package:modal_bottom_sheet/modal_bottom_sheet.dart';
 export 'package:toastification/toastification.dart';
 export 'package:flutter_svg/flutter_svg.dart';
+export 'src/ui/future_dialog/future_dialog.dart';
 export 'src/ui/fluid_dialog/fluid_dialog.dart';
 export 'src/ui/text_with_stroke.dart';
 export 'src/ui/rich_text_widget.dart';