+++ lib/src/extensions/ui_extensions.dart
@@ -1,0 +1,9 @@
+part of 'extensions.dart';
+
+extension CircularProgressExtension on CircularProgressIndicator {
+  SizedBox resized(double size) => SizedBox(
+        width: size,
+        height: size,
+        child: this,
+      );
+}