+++ lib/src/ui/future_dialog/future_dialog.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+import 'dart:async';
+
+import 'package:hayiqu/src/extensions/extensions.dart';
+
+part 'src/future_dialog_extension.dart';
+part 'src/loading_controller.dart';
+part 'src/loading_dialog.dart';
+part 'src/loading_view.dart';