+++ CHANGELOG.md
@@ -1,3 +1,12 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+Package
+
+## 0.0.2
+English translation
+
+## 0.0.3
+Added some extensions
+
+## 0.0.4
+Added loading dialog and some extensions